+++ packages/terra-time-input/tests/jest/TimeInput.test.jsx
@@ -2,7 +2,7 @@
 
 import React from 'react';
 /* eslint-disable-next-line import/no-extraneous-dependencies */
-import { shallowWithIntl, mountWithIntl, renderWithIntl } from 'terra-enzyme-intl';
+import { shallowWithIntl, mountWithIntl, renderWithIntl } from '@cerner/terra-enzyme-intl';
 import * as KeyCode from 'keycode-js';
 import ThemeContextProvider from 'terra-theme-context/lib/ThemeContextProvider';
 import TimeInput from '../../lib/TimeInput';
@@ -69,12 +69,12 @@ it('should render a time input with 12 hour clock set', () => {
 
 it('should ignore invalid times properly', () => {
   const timeInput = <TimeInput name="time-input" value="11:2" />;
-  const wrapper = shallowWithIntl(timeInput);
-  const testComponent = wrapper.dive();
+  const wrapper = mountWithIntl(timeInput);
+  const testComponent = wrapper;
   expect(testComponent).toMatchSnapshot();
 
-  expect(testComponent.instance().state.hour).toEqual('');
-  expect(testComponent.instance().state.minute).toEqual('');
+  expect(testComponent.find('TimeInput').instance().state.hour).toEqual('');
+  expect(testComponent.find('TimeInput').instance().state.minute).toEqual('');
 });
 
 it('should render a 24 hour timepicker properly on mobile devices', () => {
@@ -228,13 +228,12 @@ it('should render a 12 hour timepicker meridiem with buttons and seconds input w
 
 it('should ignore invalid times with seconds properly', () => {
   const timeInput = <TimeInput name="time-input" value="11:25:4" showSeconds />;
-  const wrapper = shallowWithIntl(timeInput);
-  const testComponent = wrapper.dive();
+  const wrapper = mountWithIntl(timeInput);
+  const testComponent = wrapper;
   expect(testComponent).toMatchSnapshot();
-
-  expect(testComponent.instance().state.hour).toEqual('');
-  expect(testComponent.instance().state.minute).toEqual('');
-  expect(testComponent.instance().state.second).toEqual('');
+  expect(testComponent.find('TimeInput').instance().state.hour).toEqual('');
+  expect(testComponent.find('TimeInput').instance().state.minute).toEqual('');
+  expect(testComponent.find('TimeInput').instance().state.second).toEqual('');
 });
 
 it('should validate the incrementHour helper method', () => {